fix: eliminate redundant function definitions
