+++ PBL5/4_Facebook_Newsfeed/backend/routes/feedRoutes.js
@@ -2,7 +2,6 @@ const router = require("express").Router()
 const { handleGetFeed } = require("../controllers/feedController")
 
 // Get feed for a specific user (posts from users they follow)
-// Example: GET /api/feed/1?limit=10&offset=0
-router.get("/:id", handleGetFeed)
+router.get("/", handleGetFeed)
 
 module.exports = router
